AVM: Reorganize the crypto opcodes a bit to simplify incentive work (#5787)
