+++ data/transactions/logic/langspec_v6.json
@@ -3,46 +3,64 @@
   "LogicSigVersion": 9,
   "NamedTypes": [
     {
-      "Name": "uint64",
-      "Abbreviation": "i",
+      "Name": "[32]byte",
+      "Abbreviation": "3",
       "Bound": [
-        0,
-        18446744073709551615
+        32,
+        32
       ],
-      "AVMType": "uint64"
+      "AVMType": "[]byte"
     },
     {
-      "Name": "stateKey",
-      "Abbreviation": "K",
+      "Name": "[64]byte",
+      "Abbreviation": "6",
       "Bound": [
-        0,
+        64,
         64
       ],
       "AVMType": "[]byte"
     },
     {
-      "Name": "none",
-      "Abbreviation": "x",
+      "Name": "[80]byte",
+      "Abbreviation": "8",
       "Bound": [
-        0,
-        0
-      ],
-      "AVMType": "none"
-    },
-    {
-      "Name": "method",
-      "Abbreviation": "M",
-      "Bound": [
-        4,
-        4
+        80,
+        80
       ],
       "AVMType": "[]byte"
     },
     {
-      "Name": "boxName",
-      "Abbreviation": "N",
+      "Name": "[]byte",
+      "Abbreviation": "b",
       "Bound": [
-        1,
+        0,
+        4096
+      ],
+      "AVMType": "[]byte"
+    },
+    {
+      "Name": "address",
+      "Abbreviation": "A",
+      "Bound": [
+        32,
+        32
+      ],
+      "AVMType": "[]byte"
+    },
+    {
+      "Name": "any",
+      "Abbreviation": "a",
+      "Bound": [
+        0,
+        0
+      ],
+      "AVMType": "any"
+    },
+    {
+      "Name": "bigint",
+      "Abbreviation": "I",
+      "Bound": [
+        0,
         64
       ],
       "AVMType": "[]byte"
@@ -57,8 +75,35 @@
       "AVMType": "uint64"
     },
     {
-      "Name": "bigint",
-      "Abbreviation": "I",
+      "Name": "boxName",
+      "Abbreviation": "N",
+      "Bound": [
+        1,
+        64
+      ],
+      "AVMType": "[]byte"
+    },
+    {
+      "Name": "method",
+      "Abbreviation": "M",
+      "Bound": [
+        4,
+        4
+      ],
+      "AVMType": "[]byte"
+    },
+    {
+      "Name": "none",
+      "Abbreviation": "x",
+      "Bound": [
+        0,
+        0
+      ],
+      "AVMType": "none"
+    },
+    {
+      "Name": "stateKey",
+      "Abbreviation": "K",
       "Bound": [
         0,
         64
@@ -66,40 +111,13 @@
       "AVMType": "[]byte"
     },
     {
-      "Name": "any",
-      "Abbreviation": "a",
+      "Name": "uint64",
+      "Abbreviation": "i",
       "Bound": [
         0,
-        0
+        18446744073709551615
       ],
-      "AVMType": "any"
-    },
-    {
-      "Name": "address",
-      "Abbreviation": "A",
-      "Bound": [
-        32,
-        32
-      ],
-      "AVMType": "[]byte"
-    },
-    {
-      "Name": "[]byte",
-      "Abbreviation": "b",
-      "Bound": [
-        0,
-        4096
-      ],
-      "AVMType": "[]byte"
-    },
-    {
-      "Name": "[32]byte",
-      "Abbreviation": "H",
-      "Bound": [
-        32,
-        32
-      ],
-      "AVMType": "[]byte"
+      "AVMType": "uint64"
     }
   ],
   "Ops": [
@@ -128,7 +146,7 @@
       "Doc": "SHA256 hash of value A, yields [32]byte",
       "IntroducedVersion": 1,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
@@ -145,7 +163,7 @@
       "Doc": "Keccak256 hash of value A, yields [32]byte",
       "IntroducedVersion": 1,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
@@ -162,7 +180,7 @@
       "Doc": "SHA512_256 hash of value A, yields [32]byte",
       "IntroducedVersion": 1,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
@@ -170,8 +188,8 @@
       "Name": "ed25519verify",
       "Args": [
         "[]byte",
-        "[]byte",
-        "[]byte"
+        "[64]byte",
+        "[32]byte"
       ],
       "Returns": [
         "bool"
@@ -182,14 +200,14 @@
       "DocExtra": "The 32 byte public key is the last element on the stack, preceded by the 64 byte signature at the second-to-last element on the stack, preceded by the data which was signed at the third-to-last element on the stack.",
       "IntroducedVersion": 1,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
       "Opcode": 5,
       "Name": "ecdsa_verify",
       "Args": [
-        "[]byte",
+        "[32]byte",
         "[]byte",
         "[]byte",
         "[]byte",
@@ -215,7 +233,7 @@
       ],
       "IntroducedVersion": 5,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
@@ -245,17 +263,17 @@
       ],
       "IntroducedVersion": 5,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
       "Opcode": 7,
       "Name": "ecdsa_pk_recover",
       "Args": [
-        "[]byte",
+        "[32]byte",
         "uint64",
-        "[]byte",
-        "[]byte"
+        "[32]byte",
+        "[32]byte"
       ],
       "Returns": [
         "[]byte",
@@ -278,7 +296,7 @@
       ],
       "IntroducedVersion": 5,
       "Groups": [
-        "Arithmetic"
+        "Cryptography"
       ]
     },
     {
@@ -531,7 +549,7 @@
       "Doc": "yields length of byte value A",
       "IntroducedVersion": 1,
       "Groups": [
-        "Arithmetic"
+        "Byte Array Manipulation"
       ]
     },
     {
@@ -2079,7 +2097,7 @@
       "DocExtra": "`concat` fails if the result would be greater than 4096 bytes.",
       "IntroducedVersion": 2,
       "Groups": [
-        "Arithmetic"
+        "Byte Array Manipulation"
       ]
     },
     {
@@ -2146,7 +2164,7 @@
       "DocExtra": "see explanation of bit ordering in setbit",
       "IntroducedVersion": 3,
       "Groups": [
-        "Arithmetic"
+        "Byte Array Manipulation"
       ]
     },
     {
@@ -2166,7 +2184,7 @@
       "DocExtra": "When A is a uint64, index 0 is the least significant bit. Setting bit 3 to 1 on the integer 0 yields 8, or 2^3. When A is a byte array, index 0 is the leftmost bit of the leftmost byte. Setting bits 0 through 11 to 1 in a 4-byte-array of 0s yields the byte array 0xfff00000. Setting bit 3 to 1 on the 1-byte-array 0x00 yields the byte array 0x10.",
       "IntroducedVersion": 3,
       "Groups": [
-        "Arithmetic"
+        "Byte Array Manipulation"
       ]
     },
     {
@@ -2184,7 +2202,7 @@
       "Doc": "Bth byte of A, as an integer. If B is greater than or equal to the array length, the program fails",
       "IntroducedVersion": 3,
       "Groups": [
-        "Arithmetic"
+        "Byte Array Manipulation"
       ]
     },
     {
@@ -2203,7 +2221,7 @@
       "Doc": "Copy of A with the Bth byte set to small integer (between 0..255) C. If B is greater than or equal to the array length, the program fails",
       "IntroducedVersion": 3,
       "Groups": [
-        "Arithmetic"
+        "Byte Array Manipulation"
       ]
     },
     {
